RingBuffer#capacity is a int, not a long
Turns out scala's type conversion was upcasting capacity to long. This
explains another dumb line I had in Writer

+++ src/main/java/RingBuffer.java
@@ -3,6 +3,6 @@ package com.somethingsimilar.ring_buffer;
 interface RingBuffer<T> {
   long add(T obj);
   T get(long slot);
-  long capacity();
+  int capacity();
   long latestSlot();
 }
